[sherlock] Add clock ID for h264 video encoder clock gate
The h264 encoder shares a clock register with the decoder.
Add clock ID and bind to it so we can synchronize it's use.

As the driver that binds the composite device is in a different repo,
the test for this is in a follow up CL. Existing device enumeration
tests ensure this doesn't break anything.

Test: fx test h264_encoder_test
Bug: 45193
Change-Id: Id4951f256ff3b1aad9541c32afa55a875afca907
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428056
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Testability-Review: Gurjant Kalsi <gkalsi@google.com>
Commit-Queue: Austin Foxley <afoxley@google.com>