[llcpp] Separate client channel/binding ownership.
This CL aims to solve the following problem:
Presently, the lifetime of the channel is a subset of the lifetime of
the AsyncBinding, thus a std::shared_ptr<AsyncBinding> is used to
guarantee safe access. The unbind operation, executing on either a
dispatcher thread or a thread shutting down the dispatcher, blocks on
the release of all references to the AsyncBinding. For a server, all
additional references are transient. For a client, this could block on a
synchronous call to another process, potentially locking up the entire
process.

This CL solves the problem by separating the channel and AsyncBinding
lifetimes and sharing ownership of it between the AsyncBinding and the
ClientBase. This manifests as a new ChannelSyncRef class. The ClientBase
and AsyncBinding (via the OnUnboundFn) each hold a
std::shared_ptr<ChannelSyncRef>. Temporary references are generated
from a std::weak_ptr<ChannelSyncRef> for the duration of any outgoing
calls. On destruction, the ChannelSyncRef can be configured to signal a
sync_completion_t, enabling a new fidl::Client API WaitForChannel()
which is used instead of the OnClientUnboundFn to recover the channel
safely. A consequence of the new API is that the channel can no longer
be recovered via the OnClientUnboundFn, thus zx::channel argument has
been removed.

NOTE: Another possibility is to simply just post a task to invoke the
OnUnboundFn from ~AsyncBinding() rather than blocking in a dispatcher
thread. However, this is problematic in the case of dispatcher shutdown,
as the OnUnboundFn would then need to be executed on the thread on which
~AsyncBinding() is running. This could be nested within other user-code,
potentially triggering a deadlock.

Bug: None.
Test: fidl-walker-tests
Change-Id: Icd592da6f0e9daaed9d46d94651d6055a9f06133
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427244
Commit-Queue: Madhav Iyengar <madhaviyengar@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Testability-Review: Andres Oportus <andresoportus@google.com>
Testability-Review: Yifei Teng <yifeit@google.com>