[usb] Support multiple configurations for usb-peripheral mode
This patch adds support for multiple configurations bound concurrently
in peripheral mode.
It also adds support for another Realtek adapter with a proprietary
interface on the first configuration and a CDC Ethernet interface
on the second configuration. A CDC Ethernet interface is used in the
virtual-bus for both configurations due to the lack of a specification
to emulate the proprietary interface.

Change-Id: I1f05bea2bf93bf327eccccb6f3ea6631ec1e56dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428434
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>
API-Review: Ruchira Ravoori <ravoorir@google.com>