[fdf] Remove <zircon/driver/binding.h>
Move the contents into the DDK and replace all references to the old
header with that. This is both a clean-up, there's no need for this
header to be in the zircon public includes, and will simplify future
changes to how drivers use bind rules.

Change-Id: I1c77381eb816c885cbf374ae8c8854a407ca504c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430319
Commit-Queue: Alex Legg <alexlegg@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Abdulla Kamar <abdulla@google.com>