[fvm] Use new Superblock getters.
Reports one less for the number of usable partition table entries than
previously. This accounts for the unused 0th entry. Comments are added
in various places to document this.

Provide default initialization for the fields of the fvm::Header. The
static assert that the struct is trivial is removed.

A new static constructor is added to fvm::Header to construct one from a
number of slices. This has the new form which will require the partition
count. In a later pass, the constructors from fvm::FormatInfo willl be
moved here also to construct from a disk size, and fvm::FormatInfo will
be deleted.

Removes the PartitionTable and AllocationTable structures which held
the expected offset and size of these tables. Use the Header getters
when possible. Some new helper functions are added to compute these when
needed. These will need some enhancement in a later pass to account for
varying partition table sizes.

A new unit test file is added for the Header getters. An underflow bug
is fixed for 0-length allocation tables that was identified by this
test. The goal is that these replace the format-info-tests when
FormatInfo is removed.

FVMDescriptor is fixed to account for two copies of the metadata when
given a target volume size. an FVMDescriptor test was removed because it
just duplicates the same equation as the implementation.

The FVM driver now checks for properly-aligned allocation table sizes
when it validates the header.

fvm::AllocTableLength is renamed to fvm::AllocTableLengthForDiskSize to
make more sense with the new variant that takes a slice count.

Change-Id: Icfb624927c14075e69283ce687e1fec261249804
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429465
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Reviewed-by: Bryan Henry <bryanhenry@google.com>