[build][alln] Add //src/zircon/lib/zircon target
This CL introduces //src/zircon/lib/zircon, a new target
that should be used by any other target that needs to link
to the Zircon VDSO (i.e. libzircon.so).

For build unification, moving the sysroot to the Fuchsia build
requires that all uses of:

  libs = [ "zircon" ]

Be replaced by a dependency to the Fuchsia-build zircon target, i.e.:

  deps += [ "//zircon/system/ulib/system" ]

However, this needs to be done atomically, and there are multiple
use cases of `libs` in third_party and vendor git repositories.

To work-around this, an intermediate target is introduced, and
all use cases in fuchsia.git are modified to use it instead
of `libs`. Note that this doesn't change the build.

Other CLs will update the targets under third_party/ and vendor/
to do the same.

Once all is done, the definition of //src/zircon/lib/zircon can
be changed to redirect to //zircon/system/ulib/zircon (or to
any other location if that target is moved to //src/zircon/
in the future).

+ Update the definition of build/config/fuzzer/BUILD.gn and
  build/fuzzing/fuzzer.gni, since it is not possible to add
  a dependency in a config, this must be done in the template
  that uses it instead.

Bug: 60613
Testing: No change to the build!

Change-Id: Ibabd5062f44f8a707766478aad496ed3e1cc3639
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433557
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Commit-Queue: David Turner <digit@google.com>