[fvm] Ensure reserved partition is paved
The reserve partition added in fxrev.dev/440062 was always being deleted
during paving, since it started out as inactive and was never made
active by a call to Upgrade.

Fix this by making the paver always upgrade the reserve partition.

Doing this required exposing the previously-hidden reserve partition as
a regular block device, since the paver makes a FIDL call to the block
device to get the partition GUID before attempting to upgrade the
reserve partition.

We also change how we detect internal reservation partitions to matching
the type, not the GUID. This is necessary because the GUID changes every
re-pave, but the type is stable. Note that this detection heuristic is
currently only used for logging, but eventually we need it to work so
that we can reclaim these slices.

Tested: Paved from before->after this change.
Tested: Paved from this change->new build with this change.

Change-Id: Ic43d69e4d3e59f4ecee99537c3c1f4f4cb968c93
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442196
Reviewed-by: Brett Wilson <brettw@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>