[fvm] Change Metadata to store just one copy
Simplifies fvm::Metadata to only store one copy of the metadata (rather
than both of the A/B copies).

Because we are no longer planning to add snapshot information in a new
fixed-offset table (instead, we'll use reserved slices), it is no longer
necessary to hide the general layout of the metadata in the
fvm::Metadata object; instead clients can ask what offset/length to
write the A and B copies to.

This is simpler for when we integrate fvm::Metadata into the driver,
since the driver currently uses this model (where there is one in-memory
copy that is written to alternating offsets).

Bug: 60709
Change-Id: If64ac718f3aeb6b8c01d4a60d5754d5f537803c3
Tested: Built and booted a variety of fvm images
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438495
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
Testability-Review: Brett Wilson <brettw@google.com>