[securemem][logspam] Lower severity on non-bug logs
On boot, sysmem attempts to communicate with the securemem TA to
configure and discover secure memory heaps. This process failing is not
necessarily a bug, such as when sysmem is used during Zedboot. The TA is
unavailable in zedboot, so the secure heap configuration will fail, but
those heaps are not needed for zedboot anyways. This CL lowers the
severity to a WARNING and tweaks some of the log statements to include
relevant info.

Change-Id: Ib4564ebfc79302b1c8f529ccafbca22c1776f9a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447540
Commit-Queue: RJ Ascani <rjascani@google.com>
Fuchsia-Auto-Submit: RJ Ascani <rjascani@google.com>
Reviewed-by: Ali Zhang <alizhang@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Testability-Review: John Bauman <jbauman@google.com>