[sdio] Add RunDiagnostics method for extra logging after client errors
There have been several instances of the brcmfmac firmware download
failing on Sherlock without triggering SDIO CRC errors. On the surface
they don't appear to be related to SDIO, but it would be helpful to have
more information on the conditions of the bus and the card when this
happens. Add a RunDiagnostics method to the SDIO protocol which can be
called when the client encounters issues such as this.

Bug: fxbug.dev/64166
Test: fx test sdmmc-test
Change-Id: I01f6264db3378c751cf6dc114669d029eeea60dc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449817
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>