[fvm] Reset partition limit for new partitions.
The partition limit is not persisted to the device, but the FVM driver
was persisting the limit in memory for the partition index.

This change resets the partition limit when a partition is created to
make sure there is no leftover data in the array.

Change-Id: If660c7a4ed482526df5276c2653870419141e7ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450460
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Testability-Review: Chris Suter <csuter@google.com>