[aml-rawnand] lower ECC errors to warnings
On raw NAND, ECC errors aren't always critical - we might be just
discovering that a block has gone bad, or we might need to retry a read
operation.

Lowering these logs to WARNINGS allows the higher-level caller to handle
it if possible, or log an ERROR itself if it really is a critical
failure.

Bug: 64858, 64862
Change-Id: Ic794913364b128d67f33c8e54b148e5dbe09e21f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/453814
Commit-Queue: David Pursell <dpursell@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>