[ftl][ndm]: Reduce bad block to debug.
This is just reporting the list of bad blocks found in the device.
This is expected, as there is a block reservation that replaces blocks
gone bad.
Essentially this is the list of bad blocks reported by the parent
device, or treated as the factory bad block.

Fixed: 66307
TEST=CQ

Change-Id: I08417ddb71ece7463ca16f6b7e5aafa35c6adf70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461523
Reviewed-by: James Sullivan <jfsulliv@google.com>
Testability-Review: James Sullivan <jfsulliv@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>