[log] Remove AmlCpu::GetPerformanceStateInfo error
Use an INFO level log instead of ERROR for a call to
AmlCpu::GetPerformanceStateInfo that supplies an out-of-bounds operating
point argument because this scenario is expected and not an error.

During initialization, power-manager will query the CPU driver to find
out the frequency/voltage information for each of the driver's supported
power states. power-manager does this by calling GetPerformanceStateInfo
for each power state. Since the power-manager doesn't know how many
power states are supported, it queries the states 0-20
(MAX_DEVICE_PERFORMANCE_STATES) until an error is returned for an
invalid state.

Test: log change only
Change-Id: Ic2bcf226ebb02cd6eec6d4b880f1071aae7af0b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/456864
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>