[storage]: Ftl Integration Test.
Fixed test that was not building anymore due to a previous CL which
added a ftl root node to the inspect hierarchy.

Added build only integration test, so it doesnt break again. This is
achieved by setting the environment as an empty list for the package.

TEST=CQ

Change-Id: If45fa527e5b39f1fdb019614b0c1bd737bf738d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467205
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>