[bindc] Add support for the new bytecode format
This CL adds an "use-new-bytecode" arg to the bindc program.
When the arg is true, the bytecode produced will be in the new
bytecode format.

The bind compiler is refactored so that the instructions
and the symbol table are wrapped in a BindProgram struct. In
addition, the logic for encoding the instructions into the
old bytecode format are moved into a separate file. A skeleton
file for the new bytecode format logic is added.

Tests: Added a unit test for argument
Bug: 67440
Change-Id: I3fab17644f2a842b6adcc276b918bd57e6ce0621
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/469675
Reviewed-by: Sarah Chan <spqchan@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>