[tee][ddk] Replace tee.Device with tee.Application
The TEE banjo protocol essentially just serves as a connector to the
fuchsia.tee.Device FIDL protocol. The fuchsia.tee.Device protocol is
being replaced with a protocol that is scoped to a single trusted
application instead of the entire TEE. In order to appropriately scope
the channel, the client must now provide the trusted application's uuid
when the connection is first established. This CL adds a new method to
the TEE ddk protocol to provide the UUID and establish a
fuchsia.tee.Application channel. It also updates the clients to use the
new protocol.

Bug: 44664
Change-Id: I7fe753d2a6c3645c65458e8e6391e3d0ac095377
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403569
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Jonathan Bailey <jonathanbailey@google.com>
Reviewed-by: RJ Ascani <rjascani@google.com>
Commit-Queue: Jonathan Bailey <jonathanbailey@google.com>